requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0

